backend/requirements.txt
fastapi
uvicorn
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart